[BUGFIX] Do not update slug when moving a locked page
Resolves: #46

With this commit one is able to move a page which has tx_sluggi_lock=1 and its slug is not updated. Also the slug of child pages are not updated.

When you move a page with tx_sluggi_lock=0 (1) whose child has tx_sluggi_lock=1 (2), then page 1 gets a new slug and page 2 slug is not updated.